refactor(media): Extract media library code in its own file
